Move communication benchmarks
