+++ benchmarks/communication/all_gather.py
@@ -4,8 +4,8 @@ import sys, os, time
 COMMS_BENCH_DIR = os.path.join(os.path.dirname(__file__), "../")
 sys.path.append(COMMS_BENCH_DIR)
 
-from .utils import *
-from .constants import *
+from communication.utils import *
+from communication.constants import *
 
 
 # Run all_gather and print metrics
