+++ benchmarks/communication/run_all.py
@@ -3,13 +3,13 @@ import sys, os
 COMMS_BENCH_DIR = os.path.join(os.path.dirname(__file__), "../")
 sys.path.append(COMMS_BENCH_DIR)
 
-from .utils import *
-from .all_reduce import run_all_reduce
-from .all_gather import run_all_gather
-from .all_to_all import run_all_to_all
-from .pt2pt import run_pt2pt
-from .broadcast import run_broadcast
-from .constants import *
+from communication.utils import *
+from communication.all_reduce import run_all_reduce
+from communication.all_gather import run_all_gather
+from communication.all_to_all import run_all_to_all
+from communication.pt2pt import run_pt2pt
+from communication.broadcast import run_broadcast
+from communication.constants import *
 
 
 # For importing
